add total_price to service method
