+++ service/serializers.py
@@ -160,6 +160,7 @@ class ServiceDetailSerializer(ServiceSerializer):
             "service_send_time",
             "service_get_time",
             "main_status",
+            "total_price",
         ]
         read_only_fields = [
             "host",
@@ -184,6 +185,7 @@ class ServiceDetailSerializer(ServiceSerializer):
             "is_customer_receive_dog",
             "created_deposit_payment",
             "created_late_payment",
+            "total_price",
             "main_status",
         ]
